refactoring and restructing
