+++ src/components/DeclensionTable.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/DeclensionTable.css";
+import "./styles/DeclensionTable.css";
 import { cleanUnderscoresToProper } from "../js/helpers";
 
 const DeclensionTable = (props) => {
@@ -11,6 +11,8 @@ const DeclensionTable = (props) => {
     nominative,
     tentativeRoot,
   } = props;
+
+  console.log(props);
 
   return (
     <table
