+++ src/components/styles/DeclensionTable.css
