+++ src/components/styles/Decliner.css
